Move tests to proper class. Fix issue with version referencing
